SDK-1212: Refactor sandbox code location to move it out of main source directory
