+++ yoti_python_sandbox/age_verification.py
@@ -1,4 +1,4 @@
-from yoti_python_sdk.sandbox.attribute import SandboxAttribute
+from yoti_python_sandbox.attribute import SandboxAttribute
 from yoti_python_sdk import config
